fix: explicitly state stacklevel=2 in warnings
The default behavior for `warnings.warn` is to only print out the location
where the warning was issued. This is usually not the most helpful behavior as
it does not tell the consumer of build123d what they did to trigger the
warning. By setting stacklevel=2 the call site will be marked with the warning
which is almost surely what is desired.